kafka replay speed: add alert for when we miss records in Kafka (#9921)
* kafka replay speed: don't trim fetchWants

I realized that trimming `fetchWant`s can end up discarding offsets in extreme circumstances.

### How it works

If the fetchWant is so big that its size would exceed 2GiB, then we trim it. We trim it by reducing the end offset. The idea is that the next fetchWant will pick up from where this one left off.

### How it can break

We trim the `fetchWant` in `UpdateBytesPerRecord` too. `UpdateBytesPerRecord` can be invoked in `concurrentFEtchers.run` after the `fetchWant` is dispatched. In that case the next `fetchWant` would have already been calculated. And we would end up with a gap.

### Did it break?

It's hard to tell, but it's very unlikely. To reach 2GiB we would have needed to have the estimation for bytes per record be 2 MiB. While these large records are possible, they should be rare and our rolling average estimation for records size shouldn't reach it.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* kafka replay speed: add alert for when we miss records in Kafka

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Restore local config

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Assert there are no missed records at the end of every test

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* make doc

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Fix rebase

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Add support for gaps within a Fetch

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Reword runbook

Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>

* Update log fields

Co-authored-by: Marco Pracucci <marco@pracucci.com>

* Update docs/sources/mimir/manage/mimir-runbooks/_index.md

* Add TestFindGapsInRecords

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>
Co-authored-by: Marco Pracucci <marco@pracucci.com>