Hash object names used as cache keys for CachingBucket
When caching rule groups, the base64 encoded name of the rule group is
used as part of the key for caching its contents. For long rule group
names, this can exceed the max key length of Memcached. To solve this
we use the sha256 of the object name in the cache key instead of the
object name itself.

Related #9386