fix(distributor-ha): Use a no-op HA tracker when not accepting writes
This changes how the HA tracker in the distributor is initialized. Instead
of creating it unconditionally when a distributor is started, use a no-op
version when the HA tracker is disabled or when running as part of component
that does not accept writes but still uses distributor code like queriers
and rulers.

Related #9826

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>