frontend: add new query pruning
The current method of excluding/including sub query results in PromQL
by comparing to -Inf or +Inf is no longer valid after
https://github.com/prometheus/prometheus/pull/15245
due to comparison of native histograms to a float with < or > result
in Jeanette's warning, not empty set.

The new method uses logical AND operation to intersect the sub query with
either a const vector() or an empty vector(). E.g.

subquery and on() (vector(1)==1)
subquery and on() (vector(-1)==1)

which become:

subquery
(vector(-1)==1)

Note that although in theory (vector(-1)==1) could be dropped in some
cases, it depends on the context and out of scope for this PR.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>