Distributor: Artificially delay sample ingestion if configured
With Sigyn, we have a much higher latency in ingestion - somewhere around 5x. Ths can produce adverse side affects on remote write clients struggling. The purpose of this change is to help remote write clients prepare for the increase in latency by artifically producing it.

The delay is meant to be configurable per tenant.

Signed-off-by: gotjosh <josue.abreu@gmail.com>