UsageTracker: don't CAS when tracking series
When we're tracking series, the timestamp we're setting is `now`
truncated to minutes. It's VERY unlikely that some other goroutine has
updated this timestamp to something newer already.

Maybe there was a newer timestamp set from a snapshot, in which case we
should overwrite it, as we cleanup based on *our* wall clock anyway.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>