Fix TestTrackerStore_Cleanup_Concurrency
One of the assumptions of trackerStore is that cleanup() is called at
least once in each idleTimeout: this is needed to make sure that
"minutes" timestamp doesn't get too old and starts looking as a newer
one.

We didn't ensure this in the test, so weird stuff was happening.

Now we increase the timestamp one minute each time, and we set idle
timeout to 1 minute, so we force the race conditions.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>