Distributor: Delay simulation on ingestion (#10107)
* Distributor: Artificially delay sample ingestion if configured

With Sigyn, we have a much higher latency in ingestion - somewhere around 5x. Ths can produce adverse side affects on remote write clients struggling. The purpose of this change is to help remote write clients prepare for the increase in latency by artifically producing it.

The delay is meant to be configurable per tenant.

Signed-off-by: gotjosh <josue.abreu@gmail.com>

* make docs

Signed-off-by: gotjosh <josue.abreu@gmail.com>

* skip docs

Signed-off-by: gotjosh <josue.abreu@gmail.com>

* rework the implementation to use a fake time and be a middleware wrapper

* fix lint

Signed-off-by: gotjosh <josue.abreu@gmail.com>

* minor style fixes

Signed-off-by: gotjosh <josue.abreu@gmail.com>

* Address review comments

Signed-off-by: gotjosh <josue.abreu@gmail.com>

* fix unit test to take into account jitter

Signed-off-by: gotjosh <josue.abreu@gmail.com>

---------

Signed-off-by: gotjosh <josue.abreu@gmail.com>