UsageTracker: use custom swiss map & channels (#10164)
* UsageTracker: use special swiss map & channels

This refactors the entire tracker.

There are two big changes here: how the map is implemented and how we
interact with it.

How the map is implemented:

We're getting rid of usual maps storing atomic.Int32: it's an overkill
to have 32bit int to store a byte, it's even more an overkill to store a
64bit pointer to that, let's not talk about map overhead.

Instead of that, we're building kind of swiss map, I based the
implementation on the code from dolthub/swiss, you can read a blog post
how that works here:
https://www.dolthub.com/blog/2023-03-28-swiss-map/

However, most of the code was replaced.

What did I do? Swiss map has a slice groups of hash suffixes, which are one
byte, to quickly find the bucket where our element is, and it also has a
slice of groups of keys. That's still a lot of overhead for our use.

What are our restrictions:
- Our keys are uint64, so great, we don't need to hash, our keys are hashes already!.
- All our keys in a single shard have the same suffix of 7 bytes (because
  we have 128 shards).

So, I changed the implementation to "index" based on the prefix, not the
suffix of the key (series hash), and since the suffix is always the
same, we could use that space for something else... Guess what? Our
values are also 7 bits, so we're storing the key AND the value in a
single uint64 field. Now each series is just uint64+uint8 in the map
implementation (plus the overallocation of the slice, of course).

I didn't just store the value in the suffix of the uint64 data, but I
also negated it, so we can figure out whether a value is a real value or
an empty value without having to check if there's a tombstone in the
index! Cleanup is now just iterating one single slice, isn't that
awesome?

Well, but how do we coordinate all of that? We can't do atomic byte
operations (can we? IDK), but also we probably don't want to. I guessed
(to be investigated whether that was true) that lots of CPU usage comes
from having to do atomic-everything.

I head that golang is good at handling hundreds of goroutines (and I
managed tens of thousands of them in the past, although not in a
high-performance env), so I decided to just run one goroutine per each
map shard.

Each map shard has its own worker which takes care of updating the
internal state: now the synchronization overhad is in pushing the data
to the workers through a channel. But once data is in the worker, a
single shard is always updated by the same goroutine.

Say we have 10k tenants in an env, that's 10K goroutines, not that
bad... I hope. We'll see.

I also considered lowering the number of shards to 64, but in order to
achieve that we need to restrict the idleTimeout to 30 minutes, so we
can fit the minutes timestamp in a 64 byte value. I think we can do
that.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

* Shutdown tenants in the benchmark

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

* Cleanup in tests too

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

* Update comments

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

* make lint: remove unused

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

* Fix tenantshard.TestMap

* go fmt

* Redefine keyMask constant

* Make safer usage of channels and slices

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

* Make events topic configurable and change it

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

* make doc reference-help

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

* Refactor events

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>

---------

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>