UsageTracker: throttle cleanup
If we cleanup all tenant's shards in a row, we may end up in a situation
where a TrackSeries request has to wait on each one of the cleanup
calls, which causes a horrible p99.

This switches to cleanup each tenant's Nth shard first, and adds an
optional min interval between shards to cover the case when there's only
one tenant.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>