querier: check size of whole frontend response before sending it (#10154)
* Add test v1

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* querier: check size of whole frontend response before sending it

### Background

Previously we'd check the size only of the body before sending a response to the query-frontend. The reason for this check is so that we don't leave it to the gRPC library. The library returns an error and doesn't tell the server anything. This results in timeouts at the query-frontend.

### Problem

Because we only check the length of the body there are edge cases where the size of the whole response (including stats & headers) exceeds the limit. The gRPC library would still refuse to send the response and the frontend would time out waiting for it.

### Is it really happening?

I discovered this in tests with very small limits. But there are probably realistic cases where this is happening: the body is just below the limit, add stats and headers, and it's over the limit.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Add CHANGELOG.md entry

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Remove unused param

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* remove unused param

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Update test

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Remove headers

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Undo changing the flag

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Include actual size and limit in error

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Cancel processor context in tests

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Increase test timeout to 1m

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Remove comment

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>