ingester: always log CPU samples for CPU limiter
I don't see a reason why we might disable this. The memory footprint is miniscule and the CPU usage is minimal too. The samples give better granularity on what event in the past 60s might have triggered this: "did CPU increase just now or has it been increasing since 30s ago"

Since I think it's harmless, we also don't need a flag for it. The flag is experimental so we can just drop it without notice.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>