Fix G306 gosec issues
> G306: Expect WriteFile permissions to be 0600 or less

The mimirtool commands should be ok. Makes using the tool a bit more restrictive, but linters gotta lint.

My understanding of the AM changes is that they're safe too because the files written are only used immediately after that to validate them and never again, so they'll never be used by another user. But if would be nice if someone from AM-maintainers can pitch in

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>