Rule Concurrency: Support self references After deployment of to dev, we are getting some of these warnings: ``` Cyclic rule dependencies detected, falling back to sequential rule evaluation ```
It turns out to be a recording rule that is self referencing (using `last_over_time` to fall back to its last value).

Whenever a rule's unevaluated dependencies consist of only itself, it is OK to be run. A caveat is that if there are two rules recording to a metric with the same name, but with different labels, the dependency tree cannot be built and concurrency is not possible