Fix instance where context cancellation error was cached
This change fixes the error mapping done in the query sharding
middleware to correctly map all possible errors from the PromQL
engine to an appropriate APIError type. This fixes an issue where
`context.Canceled` was treated as an execution error and cached
for subsequent requests.

As part of this, instrumentation for the error caching middleware
was improved and Jaeger tracing in various Docker compose setups
was fixed by pinning the image used.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>