frontend: Add option to "spin off" subqueries as actual range queries (#10460)
* frontend: Add option to "spin off" subqueries as actual range queries
Issue: https://github.com/grafana/mimir/issues/10023

This is a new feature that is completely isolated within a new middleware so it shouldn't affect current functionality of the frontend.
For safety, it requires two configurations to be enabled:
- `--query-frontend.spin-off-instant-subqueries-to-url=<url>` on the frontend. This should be set to the URL of the frontend for optimal performance. The range queries are load balanced across frontends
- `instant_queries_with_subquery_spin_off` in tenant configs. These are regexp patterns that allow us to match individual queries (or all of them). This will allow us to opt-in queries to enable the feature gradually

The feature was developed by basing myself upon the query sharding feature. The queries are mapped into either downstream queries or subqueries. Both types of queries are run and the results are fed back into prometheus' engine and the result is calculated in the frontend.

Performance impact:
The AST mapper only selects queries that are susceptible to be improved, others are just passed on to the next middleware.
For the queries that are improved, results can be up to 50x faster. When a query is selected, the worst cases I've seen are ~equal or a bit better in performance to unmodified queries.
Further tests will be done and the mapper may be improved to detect cases that aren't optimal

PromQL results impact:
None detected from all the tests I've done

* Address PR comments

* Address PR comments + fixes
- Support offsets
- Disable `@`
- Improve tests. Run each query with a different offset each time
- Add new test cases with `offset x`
- Add new test case with a long range (more than 11000 steps). It has to be split into multiple range queries
- Allow setting query path in frontend arg (instead of hardcoding `/prometheus/api/v1/query_range`)

* Make tests faster

* Address PR comments:
- Use step align code from MQE
- Fix up comments and log messages
- Return error in case of a wrong

* Make tests faster
- put the `SpinOffQueryHandler` into its own test
- do not test offset time on all queries

* Add retries to the spun-off range queries
Haven't seen any failures so far from my testing in Kubernetes, but it _will_ happen without some retries
This makes use of the regular retry middleware, so the configured retry settings will apply to these spun-off queries as well

Also, add some more tests

* Query Optimization: Align the query's end

* Add missing case of complex expressions

* Fix case where multiple range queries are merged
+ Add test that actually tests over a long range