Fix flaky `TestConcurrentFetchers_fetchSingle`
The test was failing because the control function was used up by the periodic fetch loop of the started `concurrentFetchers`. This meant that the test sometimes ran when we don't have a control function registered and didn't inject an error.

At the same time the test cases didn't retain their control functions so that they don't interfere with each other.

This change makes the test cases independent, so we can use `KeepControl`.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>