PoC: fuzz streamingpromql package
This is a PoC imlementation of a fuzz test for streaming promql just
taking the inputs and data from the two surrounding tests (it doesn't
even try functions, etc.).

This detected that the comparision function doesn't check scalars (which
I fixed), and now it fails on the `some_metric[0]@1m` query, which
returns results for Prometheus but doesn't return any for
streamingpromql.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>