Add a linter to enforce sorted lines using lint:sorted directive.
As someone once commented somewhere: we are not savages.

The lack of sorting in the modules.go modules constant definitions was bothering me big time, especially since CostAttribution was added after All.

This sorts them and adds a linter that is triggered by adding a // lint:sorted comment which would fail if lines aren't sorted.

The linter checks lines from the line containing the directive until the next white space or a line that only contains closing bracket.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>