MQE: use more accurate estimation of histogram point size (#11156)
* MQE: use more accurate estimation of histogram point size

When getting a slice of histogram points from a memory pool, use an
estimated size for each point that is closer to the actual size of
a `FloatHistogram` even though the histogram is not stored directly
in the slice, only a pointer to it. This makes tracked memory usage
closer to reality even if not perfect.

We use a fixed size for each `HPoint` instead of tracking the size
of `FloatHistograms` when they are created because float histograms
can change size during usage. This would require updating the
tracked memory usage accurately and constantly in many different
places.

* Comment phrasing

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>

---------

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>