Query Frontend: add cortex_query_samples_processed_cache_adjusted_total metric (#11164)
The metric tracks the number of samples processed to execute a query if no query results cache was enabled.

The extent now has a SamplesProcessedPerStep field. splitAndCacheMiddleware tracks partial stats of each downstream response it made. When downstreamRequest is converted to an extent, perStepStats are approximated from totalSamplesProcessed for this request, based on request range and step. It's approximated since MQE does not expose per-step sample statistics.
When extents are used to respond to a query in the partitionCacheExtents, the corresponding perStepStats are extracted and merged. The same happens when extents are merged or truncated when filtering recent cache extents.