MQE: Include chunks loaded from store-gateway in memory estimate
This increases the tracked memory consumption when chunks streamed
from store-gateways are loaded into memory as part of series iterators.
Memory consumption is decreased when the iterators are exhausted.
This decrease is optimistic since there is no guarantee that iterators
are completely consumed during a query (but they _usually_ are unless
the query is canceled or aborted).