Fix UTF-8 not allowed in Equal field for inhibition rules
This commit fixes a bug where UTF-8 characters were not allowed in the
Equal field for inhibition rules when UTF-8 support was enabled.
This meant a tenant who was using UTF-8 characters in their alerts,
and were matching these characters in their routes, silences, etc -
could not do the same with the Equal field for their inhibition rules.
The fix is taken from prometheus/alertmanager#4292 where it is
in Alertmanager v0.28.1. It has also been tested in Mimir.