dev/mimir-ingest-storage: start pods with index 0
When we run on Kubernetes, pods start at index 0. Some workloads rely on the fact that if index 1 is running, then 0 should be at least scheduled: spread-minimizing-tokens are one of them, and usage-tracker that I'm building on a different branch is another one.

This changes the indexes in the development/mimir-ingest-storage to start with 0.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>