Grafana Alertmanager: Prevent invalid config and use correct template validation on Apply (#11306)
* Grafana Alertmanager: Prevent invalid config and use correct template validation on Apply

A combination of two issues was causing a bug when attempting to use the new template
function introduced with custom webhook payloads:
 - The SetUserGrafanaConfig API endpoint was not validating templates (or receivers).
 - ApplyConfig was validating against Cloud AM template functions even though it was
    using Grafana AM template functions to create integrations.

Together, these allowed potentially invalid config to be persisted to the grafana config
bucket. This invalid config would not successfully Apply so would cause the config to
not sync until the invalid config was corrected.

Additionally, since Apply was using the wrong template validation even valid config
could cause the above sync issue if attempting to use template functions available
in Grafana AM but not in Cloud AM (such as the new webhook payload functions).

This PR:
- Does some refactoring to reduce redundant function params and email config creation.
- Adds validation to the SetUserGrafanaConfig endpoint (specifically, template and receiver).
- Adds template count and size limits as warnings during SetUserGrafanaConfig validation.
- Ensures ApplyConfig uses the correct template loader for Grafana Alertmanager.
- Adds regression tests for the described issues.

* Linting

* Address review comments