MQE: Include chunks loaded from store-gateway in memory estimate (#11453)
* MQE: Include chunks loaded from store-gateway in memory estimate

Include chunks streamed from the store-gateway that are currently in
memory in the MQE memory usage estimate. The memory usage tracker
for the current query is propagated through the read path using the
context for the current request. This change takes advantage of the
recent gRPC update that introduced a `FreeBuffer` lifecycle method.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>

* Code review changes

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>

* Update comment to clarify why we store queriers

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>

---------

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>