Add circuit breaker to not produce Kafka records at all if the context is already canceled / expired (#11768)
#### What this PR does

In this PR I'm proposing to add a circuit breaker to not produce Kafka
records at all if the context is already canceled / expired. Rationale
in the code comment.

This PR is on top of https://github.com/grafana/mimir/pull/11766.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>