MQE: always use query planner (#11762)
#### What this PR does

This PR removes the CLI option to disable the query planner when MQE is
in use. The query planner will now always be enabled.

This PR also removes all code that catered to the situation where the
query planner was disabled.

As a side effect of this, I've also removed the dependency on the query
engine in `queryStatsMiddleware` (see commit 81207ca). This was breaking
with MQE as MQE does not fully populate the `EvalStmt` returned from
`Query.Statement()` when query planning is in use, but the use of an
engine here was not necessary.

#### Which issue(s) this PR fixes or relates to

#11004

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [covered by #10067] `CHANGELOG.md` updated - the order of entries
should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If
changelog entry is not needed, please add the `changelog-not-needed`
label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.