ci: update changelog check workflow to skip backports (#11790)
#### What this PR does

This is a follow-up to https://github.com/grafana/mimir/pull/11753

I've realized that currently, the CI fails on a backporting PR. Such a
PR typically, doesn't need a dedicated changelog entry, ie. it's
upstream PR either had one, or skipped doing it explicitly.

I think, rather than asking a reviewer of the backporting PR to set the
"changelog-not-needed" label, we can just skip these PRs.

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>