OTLP: Fix conversion of non-monotonic sum metrics (#11810)
#### What this PR does

Upgrade mimir-prometheus to a version that converts non-monotonic OTel
sum metrics correctly, i.e. doesn't append a `_total` suffix. The reason
is that non-monotonic OTel sums act like gauges (in that they can go up
and down), and not counters (which only increase).

This got broken in https://github.com/grafana/mimir/pull/11607.

#### Which issue(s) this PR fixes or relates to

Fixes #11800.

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Arve Knudsen <arve.knudsen@gmail.com>