OTLP: Convert non-monotonic OTel sum metric names correctly also for metadata (#11823)
#### What this PR does

Continuation of #11810. On closer inspection, it turned out that there's
a duplicated function for converting OTel metric name to Mimir
equivalent, that is only used when generating Mimir metric metadata. Fix
the naming there too.

#### Which issue(s) this PR fixes or relates to

Fixes #11800.

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Arve Knudsen <arve.knudsen@gmail.com>