MQE: fix issue where queries containing a duplicate expression and `timestamp()` produce incorrect results (#11830)
#### What this PR does

This PR fixes an issue in common subexpression elimination where an
expression like `timestamp(foo_time) - foo_time` would produce incorrect
results.

This happens because the instant vector selector is incorrectly
deduplicated in this case: at present, the instant vector selector
operator behaves differently if it is wrapped in `timestamp()`, but the
query planner did not take this into consideration when identifying
duplicate expressions.

A future improvement would be to allow the instant vector selector
operator to return both the sample value and sample timestamp to
different consumers, but given queries like this are rare, I've opted
for the simpler approach.

#### Which issue(s) this PR fixes or relates to

#11189

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [covered by #10067] `CHANGELOG.md` updated - the order of entries
should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If
changelog entry is not needed, please add the `changelog-not-needed`
label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.