Start tracking -compactor.max-compaction-time after the initial compaction planning phase (#11834)
#### What this PR does

Start the max compaction timer only after the first planning has been
completed. This is important in case the metas syncing (or planning in
general) is slow. With the version in `main` (that starts the timer
before the first planning) we could end up in a situation where the
planning takes longer than "max compaction time" and we would never run
compactions at all for a given tenant.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Marco Pracucci <marco@pracucci.com>