Add MimirGoThreadsTooHigh alert  (#11836)
#### What this PR does

Add MimirGoThreadsTooHigh alert , firing when a Mimir instance is
running a very high number of Go threads.

I backtested it at Grafana Labs and both the warning and critical alert
would have just fired in 1 cell over the last 30d, the one where we had
the incident that pushed me to add this new alert.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>