Query-frontend: accurate tracking of samples processed from cache (#11719)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does
This PR leverages perStepStats to accurately track samples processed
from cache, without approximation.
It's a follow-up for https://github.com/grafana/mimir/pull/11164 and
https://github.com/grafana/mimir/pull/11416

#### How it works

- Added `GetStats() `and `SetStats()` MetricsReqiestInterface and its
implementations. It allows to manipulate a "stats" query param, which is
required for MQE and Prometheus Query Engine to return PerStepStats.
- This param is always set in the
[SplitAndCacheMiddleware](https://github.com/grafana/mimir/pull/11719/files#diff-655764486d839037085bd64ff1e3f08c12902660a2a083608ac67f55c07e3434R155),
so each request will have a PerStepStats gathered.
- PerStepStats are
[stored](https://github.com/grafana/mimir/pull/11719/files) in Extents
instead of approximation.
- PerStepStats from cache are
[merged](https://github.com/grafana/mimir/pull/11719/files#diff-655764486d839037085bd64ff1e3f08c12902660a2a083608ac67f55c07e3434R326)
with the PerStepStats from a downstream request and sum is reported in
the `cortex_query_samples_processed_cache_adjusted_total`
- Bunch of [integration tests
](https://github.com/grafana/mimir/pull/11719/files#diff-5d39b940451aba1e23d7e664ca574f63bbe521344e67670e9beefe9b082c348c)for
that


**Note**, that this functionality is under the
`query-frontend.cache-samples-processed-stats` feature flag, so It's
needed to enable this flag progressively for each release wave, to keep
PerStepStats collection after the release. It also affects behaviour of
the `cortex_query_samples_processed_cache_adjusted_total` metric. If
`query-frontend.cache-samples-processed-stats` is not enabled if will
report 0 samples.

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Innokentii Konstantinov <innokenty.konstantinov@grafana.com>
Co-authored-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>