MQE: add tests for common subexpression elimination for expressions with multiple children, and add metric on number of selectors eliminated (#11879)
#### What this PR does

This PR makes two changes to common subexpression elimination:

* Tests: I was worried we weren't handling the case where a expression
has multiple children correctly, so added some tests to check. Turns out
this case is OK, but I think it'd be valuable to retain so it's not
broken in the future.
* Metric: the existing
`cortex_mimir_query_engine_common_subexpression_elimination_duplication_nodes_introduced`
metric is useful for gauging the impact of CSE, but doesn't tell us how
much load on ingesters and store-gateways we've avoided. The new metric
tells us the number of selectors eliminated and therefore gives us a
better idea of the number of avoided ingester and store-gateway
requests.

#### Which issue(s) this PR fixes or relates to

#11189 

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [covered by #10067] `CHANGELOG.md` updated - the order of entries
should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If
changelog entry is not needed, please add the `changelog-not-needed`
label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.