MQE: alternative way to ensure pooled slices are not reused (#11877)
#### What this PR does

This PR is an alternative to #11741, which implements an idea from
@bboreham
[here](https://github.com/grafana/mimir/pull/11690#issuecomment-2999379812).

Compared to #11741, this avoids a need for a linter, at the cost of a
slightly less obvious behaviour and slightly uglier code.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/11690

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.