generate-otlp: run in build image (#11841)
This makes it easier to manage the dependencies of the script and
upgrade them only in a single place. The build image already contains
these dependencies. For this I also had to move the target into the
targets that run in a build image.

example run
https://github.com/grafana/mimir/actions/runs/15878851471/job/44773419103
and PR https://github.com/grafana/mimir/pull/11860 (the point is that
the PR's CI passes and that the GH action completes)

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>