Make auto-forget periods configurable (#11923)
#### What this PR does

This adds configuration flags for previously hardcoded auto-forget
unhealthy periods. Setting the new flags to 0 will disable auto-forget,
so I deprecated the previously added store-gateway flag used to disable
auto-forget in favour of the new configuration (meanwhile, both a
respected).

#### Which issue(s) this PR fixes or relates to

Fixes an issue with compactor failing to run compaction jobs because we
deploy them with a 4m timeout and auto-forget-unhealthy-periods is set
to 10, so it takes 40 min to delete an unhealthy compactor from the ring
and proceed with compactions.

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>