fix(ingester): Keep read only when any number of compactions in progress (#11890)
#### What this PR does
This pull requests extends the guardrails for ingesters leaving the
read-only state when compaction is in progress. The PR #11664 adds a
single guardrail for force compactions only. In contrast this PR extends
to any compactions in progress (idle, force). This PR provides a fix for
the situation where an ingester was in read-only while a downscale was
in progress and it started to compact because of idleness. When doing
this early compaction job, the ingesters received an upscale event in
following left read-only mode without the early compaction having
finished.

In detail the approach taken is to repurpose the ingester-wide atomic
`i.forcedCompactionInProgress (bool)` to be an atomic counter
`i.numCompactionsInProgress` for total amount of times the
`compactBlocks` method is called. Each call of this method is
incrementing before and decrementing the counter after compaction of the
collection of TSDBs hosted. The rationale for a counter instead of a
single boolean flag is that `compactBlocks` is called concurrently in
many places for different reasons, i.e.:
- Idle compaction is a timer (plus jitter) based routine
- Force compaction is called by the Flush handler directly or during
downscales by the rollout-operator.

#### Which issue(s) this PR fixes or relates to

Fixes #11850

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>