Alertmanager: Save the original config string when updating Grafana configuration (#11796)
#### What this PR does

This PR changes the handler that saves Grafana Alertmanager configurtion
to preserve the original JSON string during unmarshaling, instead of
re-marshaling the parsed Go struct before saving it.

If we do a JSON marshalling round-trip and the Grafana config contains
an upstream Alertmanager configuration, secrets,
[example](https://github.com/prometheus/alertmanager/blob/0ce3cfb962db3cbb1649d3e816a49a13c4036cd1/config/config.go#L56),
in this configuration are lost during marshalling and saved to the
storage as redacted strings.

Instead this PR makes the handler use JSON unmarshalling to validate the
configuration, and save the original JSON to the storage.

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.