Fix util.DurationWithJitter and friends when variance is 0 (#11958)
#### What this PR does

This function was returning 0 when variance was 0, but it should be the
`input`.

I checked the usages, and all of them are non-zero constants, so no
changelog entry is needed.

#### Which issue(s) this PR fixes or relates to

Fixes a bug I found in the code.

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>