Fix estimations in TestMemoryConsumptionLimit_SingleQueries
follow-up of 11683

The memory estimations have peaks during different stages of query processing when running with `-tags stringlabels` and when running without.

When running without `stringlabels`, strings appear cheaper (because the plain `ByteSize()` doesn't account for slice overhead). Because of this, the peak is during floating point processing.

When running with `stringlabels`, strings appear more expensive. Then the peak is during labels processing.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>