Add labels configuration for gateway ingress in Helm chart (#11958)
This update introduces a new `labels` field under the `gateway.ingress` section in the `values.yaml` file, allowing users to specify custom labels for the Ingress resource. The corresponding template in `gateway-ingress.yaml` has been updated to include these labels in the metadata.

- Added `labels: {}` to `values.yaml`
- Updated `gateway-ingress.yaml` to render labels if provided

This enhancement improves the flexibility of the Ingress configuration for users deploying the Mimir distributed system.

Signed-off-by: unkls <benjamin@sendinblue.com>