Move query pruning and experimental functions middlewares earlier in query-frontend pipeline (#11939)
#### What this PR does

This PR moves the query pruning and experimental functions middlewares
earlier in the query-frontend pipeline.

This has a number of benefits:

* these middlewares now run before splitting and sharding, so they are
only applied once to the entire query rather than once per split and
sharded leg
* these middlewares no longer run after sharding, which will make it
easier to migrate sharding to a MQE query planner optimisation pass in
the future
* the experimental functions middleware now runs immediately after the
error caching middleware, so queries that will be rejected are rejected
and cached before doing as much work as possible

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>