MQE: adjust behaviour of binary operations with `on (__name__)` to match Prometheus (#11956)
#### What this PR does

This PR changes the behaviour of MQE's implementation of binary
operations to match Prometheus' query engine.

Specifically, it changes MQE to match how Prometheus handles binary
operations with `on (__name__)`. Prometheus always drops the metric name
label, even if it is listed in `on`, as discussed in
https://github.com/prometheus/prometheus/issues/16631. Changing this in
Prometheus is considered a breaking change, so [I've updated the
Prometheus
documentation](https://github.com/prometheus/prometheus/pull/16716) to
explicitly state that the current behaviour is expected, and in this PR
I'm updating MQE to match.

#### Which issue(s) this PR fixes or relates to

https://github.com/prometheus/prometheus/issues/16631

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [covered by #10067] `CHANGELOG.md` updated - the order of entries
should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If
changelog entry is not needed, please add the `changelog-not-needed`
label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.