MQE: Fix estimations in TestMemoryConsumptionLimit_SingleQueries (#11963)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

follow-up of #11683

The memory estimations have peaks during different stages of query
processing when running with `-tags stringlabels` and when running
without.

When running without `stringlabels`, strings appear cheaper (because the
plain `ByteSize()` doesn't account for slice overhead). Because of this,
the peak is during floating point processing.

When running with `stringlabels`, strings appear more expensive. Then
the peak is during labels processing.

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>