Alertmanager: Check error code fron pre-notify hooks, and ignore 204.
We we're correctly checking the error code, which was seemingly fine unless the
hook returned valid JSON, in which case this was parsed as a valid payload.

Additionally, allow the hook to return 204 (No Content) to indicate it did not
perform any processing, and do not log errors in this case.