MQE: fix issue where a query with a common subexpression could return incorrect results (#11989)
#### What this PR does

This PR fixes an issue where common subexpression elimination (CSE) can
cause a query to return incorrect query results.

The issue would occur if the ring buffer used to buffer data was in a
state where the head of the buffer is at the end of the slice, and the
tail was at the beginning of the slice. In this case, calling `Remove`
for the item at the end of the slice would return the correct value, as
would the subsequent call for the item at the beginning of the slice,
but all following calls would return incorrect values.

#### Which issue(s) this PR fixes or relates to

#11189

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [covered by #10067] `CHANGELOG.md` updated - the order of entries
should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If
changelog entry is not needed, please add the `changelog-not-needed`
label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

(cherry picked from commit 8c9fffff376a6433f9d5713cdf88ee757ab4397e)