Jsonnet: remove support for the experimental read-write deployment mode (#11974)
#### What this PR does

The read-write deployment mode has been never promoted to stable (still
experimental) and we decided to remove it completely in Mimir 3.0. In
this PR I propose a first step: removing it from jsonnet.

#### Which issue(s) this PR fixes or relates to

Part of #11887

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Marco Pracucci <marco@pracucci.com>