Jsonnet: add assertion to ensure ingester ScaledObject has min and max replicas set to a value greater than 0 (#11979)
#### What this PR does

Ensure that the min and max replicas set on the ScaledObject are greater
than 0, otherwise Keda fails to create HPA and it could cause outages
during migrations if it gets unnoticed.

Manually tested:

```
Error: evaluating jsonnet: RUNTIME ERROR: the ingester-zone-a ScaledObject must have minReplicaCount > 0
	/REDACTED/ksonnet/vendor/mimir/ingest-storage-ingester-autoscaling.libsonnet:(203:1)-(204:113)	
	Checking object assertions	
	During manifestation	

Error: evaluating jsonnet: RUNTIME ERROR: the ingester-zone-a ScaledObject must have maxReplicaCount > 0
	/REDACTED/ksonnet/vendor/mimir/ingest-storage-ingester-autoscaling.libsonnet:(206:1)-(207:113)	
	Checking object assertions	
	During manifestation	

```

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>